+++ aprs_tt.c
@@ -1561,7 +1561,7 @@ static void raw_tt_data_to_app (int chan, char *msg)
 	  alevel.mark = -2;
 	  alevel.space = -2;
 
-	  dlq_append (DLQ_REC_FRAME, chan, -1, pp, alevel, RETRY_NONE, "tt");
+	  dlq_append (DLQ_REC_FRAME, chan, -1, 0, pp, alevel, RETRY_NONE, "tt");
 	}
 	else {
 	  text_color_set(DW_COLOR_ERROR);