+++ demod_9600.c
@@ -145,6 +145,7 @@ void demod_9600_init (int samples_per_sec, int baud, struct demodulator_state_s 
 	int j;
 
 	memset (D, 0, sizeof(struct demodulator_state_s));
+	D->num_slicers = 1;
 
 	//dw_printf ("demod_9600_init(rate=%d, baud=%d, D ptr)\n", samples_per_sec, baud);
 
@@ -259,7 +260,7 @@ void demod_9600_init (int samples_per_sec, int baud, struct demodulator_state_s 
  *
  *--------------------------------------------------------------------*/
 
-static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator_state_s *D);
+static void inline nudge_pll (int chan, int subchan, int slice, int demod_data, struct demodulator_state_s *D);
 
 __attribute__((hot))
 void demod_9600_process_sample (int chan, int sam, struct demodulator_state_s *D)
@@ -377,19 +378,16 @@ void demod_9600_process_sample (int chan, int sam, struct demodulator_state_s *D
 	  /* AGC should generally keep this around -1 to +1 range. */
 
 	  demod_data = demod_out > 0;
-
-	  nudge_pll (chan, subchan, demod_data, D);
+	  nudge_pll (chan, subchan, 0, demod_data, D);
 	}
 	else {
-	  int s;
-
-	  assert (subchan == 0);
+	  int slice;
 
 	  /* Multiple slicers each feeding its own HDLC decoder. */
 
-	  for (s=0; s<D->num_slicers; s++) {
-	    demod_data = demod_out > slice_point[s];
-	    nudge_pll (chan, s, demod_data, D);		
+	  for (slice=0; slice<D->num_slicers; slice++) {
+	    demod_data = demod_out > slice_point[slice];
+	    nudge_pll (chan, subchan, slice, demod_data, D);
 	  }
 	}
 
@@ -397,7 +395,7 @@ void demod_9600_process_sample (int chan, int sam, struct demodulator_state_s *D
 
 
 __attribute__((hot))
-static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator_state_s *D)
+static void inline nudge_pll (int chan, int subchan, int slice, int demod_data, struct demodulator_state_s *D)
 {
 
 /*
@@ -425,10 +423,11 @@ static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator
  * This was optimized for 1200 baud AFSK.  There might be some opportunity
  * for improvement here.
  */
-	D->slicer[subchan].prev_d_c_pll = D->slicer[subchan].data_clock_pll;
-	D->slicer[subchan].data_clock_pll += D->pll_step_per_sample;
 
-	if (D->slicer[subchan].data_clock_pll < 0 && D->slicer[subchan].prev_d_c_pll > 0) {
+	D->slicer[slice].prev_d_c_pll = D->slicer[slice].data_clock_pll;
+	D->slicer[slice].data_clock_pll += D->pll_step_per_sample;
+
+	if (D->slicer[slice].data_clock_pll < 0 && D->slicer[slice].prev_d_c_pll > 0) {
 
 	  /* Overflow. */
 
@@ -443,20 +442,20 @@ static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator
 	  // Warning: 'descram' set but not used.
 	  // It's used in conditional debug code below.
 	  // descram =
-	  descramble (demod_data, &(D->slicer[subchan].lfsr));
+	  descramble (demod_data, &(D->slicer[slice].lfsr));
 
-	  hdlc_rec_bit (chan, subchan, demod_data, 1, D->slicer[subchan].lfsr);
+	  hdlc_rec_bit (chan, subchan, slice, demod_data, 1, D->slicer[slice].lfsr);
 	}
 
-        if (demod_data != D->slicer[subchan].prev_demod_data) {
+        if (demod_data != D->slicer[slice].prev_demod_data) {
 
 	  // Note:  Test for this demodulator, not overall for channel.
 
-	  if (hdlc_rec_gathering (chan, subchan)) {
-	    D->slicer[subchan].data_clock_pll = (int)(D->slicer[subchan].data_clock_pll * D->pll_locked_inertia);
+	  if (hdlc_rec_gathering (chan, subchan, slice)) {
+	    D->slicer[slice].data_clock_pll = (int)(D->slicer[slice].data_clock_pll * D->pll_locked_inertia);
 	  }
 	  else {
-	    D->slicer[subchan].data_clock_pll = (int)(D->slicer[subchan].data_clock_pll * D->pll_searching_inertia);
+	    D->slicer[slice].data_clock_pll = (int)(D->slicer[slice].data_clock_pll * D->pll_searching_inertia);
 	  }
 	}
 
@@ -464,7 +463,7 @@ static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator
 #if DEBUG5
 
 	//if (chan == 0) {
-	if (hdlc_rec_gathering (chan,subchan)) {
+	if (hdlc_rec_gathering (chan,subchan,slice)) {
 	
 	  char fname[30];
 
@@ -507,7 +506,7 @@ static void nudge_pll (int chan, int subchan, int demod_data, struct demodulator
  * Remember demodulator output (pre-descrambling) so we can compare next time
  * for the DPLL sync.
  */
-	D->slicer[subchan].prev_demod_data = demod_data;
+	D->slicer[slice].prev_demod_data = demod_data;
 
 } /* end nudge_pll */